modularizing transaction type
