+++ packages/transaction-controller/src/transaction-type.test.ts
@@ -1,0 +1,241 @@
+import { determineTransactionType } from './transaction-type';
+import { TransactionType } from './types';
+
+describe('determineTransactionType', () => {
+  const FROM_MOCK = '0x9e';
+  const txParams = {
+    to: '0x9e673399f795D01116e9A8B2dD2F156705131ee9',
+    data: '0xa9059cbb0000000000000000000000002f318C334780961FB129D2a6c30D0763d9a5C970000000000000000000000000000000000000000000000000000000000000000a',
+    from: FROM_MOCK,
+  };
+
+  it('returns a token transfer type when the recipient is a contract, there is no value passed, and data is for the respective method call', async () => {
+    class EthQuery {
+      getCode(_to: any, cb: any) {
+        cb(null, '0xab');
+      }
+    }
+    const result = await determineTransactionType(
+      {
+        to: '0x9e673399f795D01116e9A8B2dD2F156705131ee9',
+        data: '0xa9059cbb0000000000000000000000002f318C334780961FB129D2a6c30D0763d9a5C970000000000000000000000000000000000000000000000000000000000000000a',
+        from: FROM_MOCK,
+      },
+      // @ts-expect-error Mock eth query does not fulfill type requirements
+      new EthQuery(),
+    );
+
+    expect(result).toMatchObject({
+      type: TransactionType.tokenMethodTransfer,
+      getCodeResponse: '0xab',
+    });
+  });
+
+  it(
+    'does NOT return a token transfer type and instead returns contract interaction' +
+      ' when the recipient is a contract, the data matches the respective method call, but there is a value passed',
+    async () => {
+      class EthQuery {
+        getCode(_to: any, cb: any) {
+          cb(null, '0xab');
+        }
+      }
+      const resultWithEmptyValue = await determineTransactionType(
+        txParams,
+        // @ts-expect-error Mock eth query does not fulfill type requirements
+        new EthQuery(),
+      );
+      expect(resultWithEmptyValue).toMatchObject({
+        type: TransactionType.tokenMethodTransfer,
+        getCodeResponse: '0xab',
+      });
+
+      const resultWithEmptyValue2 = await determineTransactionType(
+        {
+          value: '0x0000',
+          ...txParams,
+        },
+        // @ts-expect-error Mock eth query does not fulfill type requirements
+        new EthQuery(),
+      );
+
+      expect(resultWithEmptyValue2).toMatchObject({
+        type: TransactionType.tokenMethodTransfer,
+        getCodeResponse: '0xab',
+      });
+
+      const resultWithValue = await determineTransactionType(
+        {
+          value: '0x12345',
+          ...txParams,
+        },
+        // @ts-expect-error Mock eth query does not fulfill type requirements
+        new EthQuery(),
+      );
+      expect(resultWithValue).toMatchObject({
+        type: TransactionType.contractInteraction,
+        getCodeResponse: '0xab',
+      });
+    },
+  );
+
+  it('does NOT return a token transfer type when the recipient is not a contract but the data matches the respective method call', async () => {
+    class EthQuery {
+      getCode(_to: any, cb: any) {
+        cb(null, '0x');
+      }
+    }
+    const result = await determineTransactionType(
+      txParams,
+      // @ts-expect-error Mock eth query does not fulfill type requirements
+      new EthQuery(),
+    );
+    expect(result).toMatchObject({
+      type: TransactionType.simpleSend,
+      getCodeResponse: '0x',
+    });
+  });
+
+  it('returns a token approve type when the recipient is a contract and data is for the respective method call', async () => {
+    class EthQuery {
+      getCode(_to: any, cb: any) {
+        cb(null, '0xab');
+      }
+    }
+    const result = await determineTransactionType(
+      {
+        ...txParams,
+        data: '0x095ea7b30000000000000000000000002f318C334780961FB129D2a6c30D0763d9a5C9700000000000000000000000000000000000000000000000000000000000000005',
+      },
+      // @ts-expect-error Mock eth query does not fulfill type requirements
+      new EthQuery(),
+    );
+    expect(result).toMatchObject({
+      type: TransactionType.tokenMethodApprove,
+      getCodeResponse: '0xab',
+    });
+  });
+
+  it('returns a contract deployment type when "to" is falsy and there is data', async () => {
+    class EthQuery {
+      getCode(_to: any, cb: any) {
+        cb(null, '');
+      }
+    }
+    const result = await determineTransactionType(
+      {
+        ...txParams,
+        to: '',
+        data: '0xabd',
+      },
+      // @ts-expect-error Mock eth query does not fulfill type requirements
+      new EthQuery(),
+    );
+    expect(result).toMatchObject({
+      type: TransactionType.deployContract,
+      getCodeResponse: undefined,
+    });
+  });
+
+  it('returns a simple send type with a 0x getCodeResponse when there is data, but the "to" address is not a contract address', async () => {
+    class EthQuery {
+      getCode(_to: any, cb: any) {
+        cb(null, '0x');
+      }
+    }
+    const result = await determineTransactionType(
+      {
+        ...txParams,
+        data: '0xabd',
+      },
+      // @ts-expect-error Mock eth query does not fulfill type requirements
+      new EthQuery(),
+    );
+    expect(result).toMatchObject({
+      type: TransactionType.simpleSend,
+      getCodeResponse: '0x',
+    });
+  });
+
+  it('returns a simple send type with a null getCodeResponse when "to" is truthy and there is data, but getCode returns an error', async () => {
+    class EthQuery {
+      getCode(_to: any, cb: any) {
+        cb(new Error('Some error'));
+      }
+    }
+    const result = await determineTransactionType(
+      {
+        ...txParams,
+        data: '0xabd',
+      },
+      // @ts-expect-error Mock eth query does not fulfill type requirements
+      new EthQuery(),
+    );
+    expect(result).toMatchObject({
+      type: TransactionType.simpleSend,
+      getCodeResponse: null,
+    });
+  });
+
+  it('returns a contract interaction type with the correct getCodeResponse when "to" is truthy and there is data, and it is not a token transaction', async () => {
+    class EthQuery {
+      getCode(_to: any, cb: any) {
+        cb(null, '0xa');
+      }
+    }
+    const result = await determineTransactionType(
+      {
+        ...txParams,
+        data: 'abd',
+      },
+      // @ts-expect-error Mock eth query does not fulfill type requirements
+      new EthQuery(),
+    );
+    expect(result).toMatchObject({
+      type: TransactionType.contractInteraction,
+      getCodeResponse: '0xa',
+    });
+  });
+
+  it('returns a contract interaction type with the correct getCodeResponse when "to" is a contract address and data is falsy', async () => {
+    class EthQuery {
+      getCode(_to: any, cb: any) {
+        cb(null, '0xa');
+      }
+    }
+    const result = await determineTransactionType(
+      {
+        ...txParams,
+        data: '',
+      },
+      // @ts-expect-error Mock eth query does not fulfill type requirements
+      new EthQuery(),
+    );
+    expect(result).toMatchObject({
+      type: TransactionType.contractInteraction,
+      getCodeResponse: '0xa',
+    });
+  });
+
+  it('returns contractInteraction for send with approve', async () => {
+    class EthQuery {
+      getCode(_to: any, cb: any) {
+        cb(null, '0xa');
+      }
+    }
+
+    const result = await determineTransactionType(
+      {
+        ...txParams,
+        value: '0x5af3107a4000',
+        data: '0x095ea7b30000000000000000000000002f318C334780961FB129D2a6c30D0763d9a5C9700000000000000000000000000000000000000000000000000000000000000005',
+      },
+      // @ts-expect-error Mock eth query does not fulfill type requirements
+      new EthQuery(),
+    );
+    expect(result).toMatchObject({
+      type: TransactionType.contractInteraction,
+      getCodeResponse: '0xa',
+    });
+  });
+});
